SystemUI: move keyguard visualizer behind notifications
Also, handle hiding and showing the visualizer when dragging
notifications, or tapping on them.
