+++ packages/SystemUI/res/layout/keyguard_bottom_area.xml
@@ -68,14 +68,4 @@
         android:tint="#ffffffff"
         android:contentDescription="@string/accessibility_unlock_button" />
 
-    <com.pheelicks.visualizer.VisualizerView
-        android:id="@+id/visualizerView"
-        android:gravity="bottom"
-        android:layout_gravity="bottom"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:clipChildren="false"
-        android:clipToPadding="false"
-        android:visibility="visible"/>
-
 </com.android.systemui.statusbar.phone.KeyguardBottomAreaView>
