+++ packages/SystemUI/src/com/android/systemui/statusbar/phone/NotificationPanelView.java
@@ -1065,7 +1065,7 @@ public class NotificationPanelView extends PanelView implements
     private void setKeyguardBottomAreaVisibility(int statusBarState,
             boolean goingToFullShade) {
         if (goingToFullShade) {
-            mKeyguardBottomArea.requestVisualizer(false, 0);
+            mStatusBar.requestVisualizer(false, 0);
             mKeyguardBottomArea.animate().cancel();
             mKeyguardBottomArea.animate()
                     .alpha(0f)
@@ -1104,7 +1104,7 @@ public class NotificationPanelView extends PanelView implements
                         .setDuration(mStatusBar.getKeyguardFadingAwayDuration()/2)
                         .start();
             }
-            mKeyguardBottomArea.requestVisualizer(false, 0);
+            mStatusBar.requestVisualizer(false, 0);
         } else if (mStatusBarState == StatusBarState.SHADE_LOCKED
                 && statusBarState == StatusBarState.KEYGUARD) {
             mKeyguardStatusView.animate().cancel();
@@ -1117,7 +1117,7 @@ public class NotificationPanelView extends PanelView implements
                     .setDuration(320)
                     .setInterpolator(PhoneStatusBar.ALPHA_IN)
                     .withEndAction(mAnimateKeyguardStatusViewVisibleEndRunnable);
-            mKeyguardBottomArea.requestVisualizer(true, 320);
+            mStatusBar.requestVisualizer(true, 320);
         } else if (statusBarState == StatusBarState.KEYGUARD) {
             mKeyguardStatusView.animate().cancel();
             mKeyguardStatusViewAnimating = false;
@@ -1854,7 +1854,7 @@ public class NotificationPanelView extends PanelView implements
     @Override
     public void onSwipingAnimationFinished(boolean snappingBack) {
         if (snappingBack) {
-            mKeyguardBottomArea.requestVisualizer(true, 300);
+            mStatusBar.requestVisualizer(true, 300);
         }
     }
 
@@ -1974,7 +1974,7 @@ public class NotificationPanelView extends PanelView implements
 
     public void setEmptyDragAmount(float amount) {
         if (amount == 0 && mStatusBarState == StatusBarState.KEYGUARD) {
-            mKeyguardBottomArea.requestVisualizer(true, 300);
+            mStatusBar.requestVisualizer(true, 300);
         }
         float factor = 0.8f;
         if (mNotificationStackScroller.getNotGoneChildCount() > 0) {
