+++ packages/SystemUI/src/com/android/systemui/statusbar/phone/PanelView.java
@@ -845,7 +845,7 @@ public abstract class PanelView extends FrameLayout {
      * Phase 1: Move everything upwards.
      */
     private void startUnlockHintAnimationPhase1(final Runnable onAnimationFinished) {
-        mKeyguardBottomArea.requestVisualizer(false, 0);
+        mStatusBar.requestVisualizer(false, 0);
         float target = Math.max(0, getMaxPanelHeight() - mHintDistance);
         ValueAnimator animator = createHeightAnimator(target);
         animator.setDuration(250);
@@ -863,7 +863,7 @@ public abstract class PanelView extends FrameLayout {
                 if (mCancelled) {
                     mHeightAnimator = null;
                     onAnimationFinished.run();
-                    mKeyguardBottomArea.requestVisualizer(true, 250);
+                    mStatusBar.requestVisualizer(true, 250);
                 } else {
                     startUnlockHintAnimationPhase2(onAnimationFinished);
                 }
@@ -900,7 +900,7 @@ public abstract class PanelView extends FrameLayout {
             public void onAnimationEnd(Animator animation) {
                 mHeightAnimator = null;
                 onAnimationFinished.run();
-                mKeyguardBottomArea.requestVisualizer(true, 200);
+                mStatusBar.requestVisualizer(true, 200);
             }
         });
         animator.start();
