+++ packages/SystemUI/src/com/android/systemui/statusbar/phone/PhoneStatusBar.java
@@ -1050,6 +1050,7 @@ public class PhoneStatusBar extends BaseStatusBar implements DemoMode,
         mExpandedContents = mStackScroller;
 
         mBackdrop = (BackDropView) mStatusBarWindowContent.findViewById(R.id.backdrop);
+        mBackdrop.setService(this);
         mBackdropFront = (ImageView) mBackdrop.findViewById(R.id.backdrop_front);
         mBackdropBack = (ImageView) mBackdrop.findViewById(R.id.backdrop_back);
 
@@ -5175,6 +5176,10 @@ public class PhoneStatusBar extends BaseStatusBar implements DemoMode,
 
     public boolean hasActiveNotifications() {
         return !mNotificationData.getActiveNotifications().isEmpty();
+    }
+
+    public void requestVisualizer(boolean show, int delay) {
+        mBackdrop.requestVisualizer(show, delay);
     }
 
     public void wakeUpIfDozing(long time, MotionEvent event) {
