+++ packages/SystemUI/src/com/android/systemui/statusbar/phone/StatusBarWindowView.java
@@ -200,11 +200,11 @@ public class StatusBarWindowView extends FrameLayout {
         if (mService.getBarState() == StatusBarState.KEYGUARD) {
             switch (ev.getAction()) {
                 case MotionEvent.ACTION_DOWN:
-                    mService.mKeyguardBottomArea.requestVisualizer(false, 0);
+                    mService.requestVisualizer(false, 0);
                     break;
                 case MotionEvent.ACTION_UP:
                 case MotionEvent.ACTION_CANCEL:
-                    mService.mKeyguardBottomArea.requestVisualizer(true, 500);
+                    mService.requestVisualizer(true, 500);
                     break;
             }
         }
