+++ packages/SystemUI/src/com/android/systemui/statusbar/stack/NotificationStackScrollLayout.java
@@ -566,10 +566,12 @@ public class NotificationStackScrollLayout extends ViewGroup
             // We start the swipe and snap back in the same frame, we don't want any animation
             mDragAnimPendingChildren.remove(animView);
         }
+        mPhoneStatusBar.requestVisualizer(true, 300);
     }
 
     @Override
     public boolean updateSwipeProgress(View animView, boolean dismissable, float swipeProgress) {
+        mPhoneStatusBar.requestVisualizer(false, 0);
         return false;
     }
 
@@ -2155,6 +2157,12 @@ public class NotificationStackScrollLayout extends ViewGroup
             mNeedsAnimation =  true;
         }
         requestChildrenUpdate();
+        if (activatedChild != null) {
+            mPhoneStatusBar.requestVisualizer(false, 0);
+        } else {
+            mPhoneStatusBar.requestVisualizer(true, 300);
+        }
+
     }
 
     public ActivatableNotificationView getActivatedChild() {
